splitting up huge direct_sampling module
