+++ examples/13_mps/00_simple_unconditional.py
@@ -39,7 +39,9 @@ print(ti)
 #   (smaller is faster, slightly noisier).
 # * ``threshold=0.0`` — the recommended DSBC mode: always take the best match.
 
-ds = gs.DirectSampling(ti, n_neighbors=12, scan_fraction=0.3, threshold=0.0)
+ds = gs.DirectSampling(
+    gs.MPSModel(ti, n_neighbors=12, scan_fraction=0.3, threshold=0.0)
+)
 field = ds([np.arange(40, dtype=float)] * 2, seed=20250616)
 
 ###############################################################################
