+++ src/gstools/mps/distance.py
@@ -9,11 +9,6 @@ import numpy as np
 
 __all__ = [
     "compute_node_weights",
-    "categorical_dist",
-    "l1_dist",
-    "l2_dist",
-    "lp_dist",
-    "variation_dist",
     "vec_categorical_dist",
     "vec_l1_dist",
     "vec_l2_dist",
@@ -80,137 +75,6 @@ def compute_node_weights(
         # (‖h‖ >= 1) — fall back to uniform rather than emit NaNs.
         return np.full(n, 1.0 / n, dtype=np.float64)
     return raw_w / total
-
-
-def categorical_dist(data_event_sim, data_event_ti, node_weights):
-    """Weighted categorical distance (Mariethoz2010 Eq. 3).
-
-    Parameters
-    ----------
-    data_event_sim : numpy.ndarray, shape (n,)
-    data_event_ti : numpy.ndarray, shape (n,)
-    node_weights : numpy.ndarray, shape (n,)
-        Normalized spatial and conditioning weights.
-
-    Returns
-    -------
-    float
-        Distance in [0, 1].
-    """
-    return float(
-        np.dot(
-            node_weights,
-            (data_event_sim != data_event_ti).astype(np.float64),
-        )
-    )
-
-
-def l1_dist(data_event_sim, data_event_ti, node_weights, d_max):
-    """Weighted L1 distance / Manhattan (Mariethoz2010 Eq. 6).
-
-    Parameters
-    ----------
-    data_event_sim : numpy.ndarray, shape (n,)
-    data_event_ti : numpy.ndarray, shape (n,)
-    node_weights : numpy.ndarray, shape (n,)
-        Normalized spatial and conditioning weights.
-    d_max : float
-        Data range for normalization.
-
-    Returns
-    -------
-    float
-        Distance in [0, 1].
-    """
-    return float(
-        np.dot(node_weights, np.abs(data_event_sim - data_event_ti) / d_max)
-    )
-
-
-def l2_dist(data_event_sim, data_event_ti, node_weights, d_max):
-    """Weighted L2 / RMS distance (Mariethoz2010 Eq. 4–5).
-
-    Parameters
-    ----------
-    data_event_sim : numpy.ndarray, shape (n,)
-    data_event_ti : numpy.ndarray, shape (n,)
-    node_weights : numpy.ndarray, shape (n,)
-        Normalized spatial and conditioning weights.
-    d_max : float
-        Data range for normalization.
-
-    Returns
-    -------
-    float
-        Distance in [0, 1].
-    """
-    return float(
-        np.sqrt(
-            np.dot(
-                node_weights,
-                ((data_event_sim - data_event_ti) / d_max) ** 2,
-            )
-        )
-    )
-
-
-def lp_dist(data_event_sim, data_event_ti, node_weights, d_max, p):
-    """Weighted Lp (Minkowski) distance.
-
-    Warning: Computationally heavier than l1_dist or l2_dist due to
-    the generic C-level pow() evaluation. Use only when p != 1.0 or 2.0.
-
-    Parameters
-    ----------
-    data_event_sim : numpy.ndarray, shape (n,)
-    data_event_ti : numpy.ndarray, shape (n,)
-    node_weights : numpy.ndarray, shape (n,)
-        Normalized spatial and conditioning weights.
-    d_max : float
-        Data range for normalization.
-    p : float
-        The Minkowski exponent (e.g., 1.5, 3.0, 5.0).
-
-    Returns
-    -------
-    float
-        Distance in [0, 1].
-    """
-    diffs = np.abs(data_event_sim - data_event_ti) / d_max
-    return float(np.sum(node_weights * (diffs**p)) ** (1.0 / p))
-
-
-def variation_dist(data_event_sim, data_event_ti, node_weights, d_max, p=2.0):
-    """Weighted variation distance (Mariethoz2010 Eq. 9, de-meaned).
-
-    Parameters
-    ----------
-    data_event_sim : numpy.ndarray, shape (n,)
-    data_event_ti : numpy.ndarray, shape (n,)
-    node_weights : numpy.ndarray, shape (n,)
-        Normalized spatial and conditioning weights.
-    d_max : float
-        Data range for normalization.
-    p : float, optional
-        Lp aggregation exponent. Default ``2.0`` (RMS, Mariethoz2010 Eq. 9).
-
-    Returns
-    -------
-    float
-        Distance in [0, 1].
-    """
-    diffs = (data_event_sim - data_event_sim.mean()) - (
-        data_event_ti - data_event_ti.mean()
-    )
-    # 2*d_max normalises the common case to [0, 1]; SG values are not bounded
-    # by the TI range (conditioning data / accumulated mean-shifts), so clamp.
-    return float(
-        min(
-            1.0,
-            np.dot(node_weights, np.abs(diffs / (2 * d_max)) ** p)
-            ** (1.0 / p),
-        )
-    )
 
 
 # ---------------------------------------------------------------------------
